This commit was manufactured by cvs2svn to create tag 'joe-3_1jupp4'.

+++ uerror.c
@@ -238,7 +238,7 @@ BW *find_a_good_bw(B *b)
 int parserrb(B *b)
 {
 	BW *bw;
-	int n;
+	long n;
 	errbuf = b;
 	freeall();
 	n = parserr(b);
@@ -253,7 +253,7 @@ int parserrb(B *b)
 
 int uparserr(BW *bw)
 {
-	int n;
+	long n;
 	errbuf = bw->b;
 	freeall();
 	n = parserr(bw->b);
